Move options to file
